Add files via upload
The program is mainly for men as the algorithm used was designed for the use of men as it incorporates facts that tie around male biology. The program takes into account your weight (kg), height(cm), and age(years) and calculates how many calories you should in-take daily based on your daily/weekly physical activity assuming the user is honest about their activity.